requirements.txt
flask==2.0.1
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-limiter==2.4.0
python-dotenv==0.19.0
gunicorn==20.1.0
requests==2.26.0
wtforms==3.0.1
email-validator==1.1.3